+++ R/dim.R
@@ -1,5 +1,11 @@
 #' Dimensions of an experiment
 #'
+#' @description
+#' `r lifecycle::badge("deprecated")`
+#'
+#' These methods were deprecated with `experiment()`. Use the corresponding
+#' base dimension operations on a `SummarizedExperiment` instead.
+#'
 #' Retrieve the dimensions of an experiment object,
 #' i.e. the number of variables and samples.
 #'
@@ -9,8 +15,10 @@
 #' @return A vector with two elements: the number of variables and the number of samples.
 #' @examples
 #' dim(real_experiment)
+#' @keywords internal
 #' @export
 dim.glyexp_experiment <- function(x) {
+  .deprecate_experiment("dim.glyexp_experiment()", "dim()")
   dim(x$expr_mat)
 }
 
@@ -18,5 +26,9 @@ dim.glyexp_experiment <- function(x) {
 #' @rdname dim.glyexp_experiment
 #' @export
 `dim<-.glyexp_experiment` <- function(x, value) {
+  .deprecate_experiment(
+    "experiment()",
+    details = "Use base dimension operations on a SummarizedExperiment instead."
+  )
   cli::cli_abort("Dimensions of an experiment could not be set manually.")
 }