+++ R/rownames-colnames.R
@@ -1,5 +1,11 @@
 #' Dimname for experiment
 #'
+#' @description
+#' `r lifecycle::badge("deprecated")`
+#'
+#' This method was deprecated with `experiment()`. Use `dimnames()` on a
+#' `SummarizedExperiment` instead.
+#'
 #' The dimnames method for [experiment()] objects are
 #' the dimnames of their expression matrix.
 #'
@@ -11,9 +17,11 @@
 #' @examples
 #' dimnames(real_experiment)
 #'
+#' @keywords internal
 #' @export
 #' @method dimnames glyexp_experiment
 dimnames.glyexp_experiment <- function(x, ...) {
-  stopifnot(is_experiment(x))
+  .deprecate_experiment("dimnames.glyexp_experiment()", "dimnames()")
+  stopifnot(.is_experiment(x))
   dimnames(x$expr_mat)
 }